move class files into OpenCage subdirectory
